Fix crashpad_snapshot_test ProcessReaderMac.* for macOS ≥ 15.4
Starting in dyld-1284.13 (macOS 15.4), _dyld_get_image_name (in
dyld4::PrebuiltLoader::path) returns an absolute path for a main
executable, even when the image was not loaded from an absolute path.
This can break test expectations when the test executable is not invoked
from an absolute path. As a workaround, just use the main executable’s
basename for comparison purposes.

Change-Id: I3846d237a7a15e886cf15cd27146556103c22c2e
Reviewed-on: https://chromium-review.googlesource.com/c/crashpad/crashpad/+/6794725
Reviewed-by: Joshua Peraza <jperaza@chromium.org>
Commit-Queue: Mark Mentovai <mark@chromium.org>