util: DO an arm64 TODO in MacModelAndBoard
The board_id returned on mac-arm64 was taken from the target-type
property of the IOPlatformExpertDevice obtained through IOKit when this
code was written on a mac-arm64 Developer Transition Kit, with a TODO to
verify with production hardware. Several generations of production
hardware have shown that this is a viable strategy. In fact, a slightly
more detailed ID can be obtained from the target-sub-type property.

mac-x86_64 continues to use the board-id property for this purpose.

Test: crashpad_util_test MacUtil.MacModelAndBoard
Change-Id: I81d5052d1a9cd7a76b6b4593fb4115a251409de7
Reviewed-on: https://chromium-review.googlesource.com/c/crashpad/crashpad/+/6798247
Commit-Queue: Mark Mentovai <mark@chromium.org>
Reviewed-by: Justin Cohen <justincohen@chromium.org>