snapshot: prepare MachOImageAnnotationsReader for macOS/iOS 26
macOS/iOS 26 introduce a new version of the crashreporter_annotations_t
structure used in __DATA,__crash_info or __DATA_DIRTY,__crash_info:
version 7. This is expected to be the value of
CRASHREPORTER_ANNOTATIONS_VERSION in the private
<CrashReporterClient.h>. Version 7 crashreporter_annotations_t
structures are empirically observed to be 328 bytes long, which is 264
bytes (or 33 uint64_ts) longer than the previously known version 5
structure. Version 5 was used from OS X 10.11 through macOS 15, and
there does not appear to be a version 6. The precise layout and meaning
of the new fields in the version 7 extension is not yet known.

Test: crashpad_snapshot_test MachOImageAnnotations.Crash*
Bug: 425331081
Change-Id: I1bc06a4a0f07c43830701e335da479f6fb5b5846
Reviewed-on: https://chromium-review.googlesource.com/c/crashpad/crashpad/+/6796433
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Commit-Queue: Mark Mentovai <mark@chromium.org>