rate-limit: create ShouldRateLimit() and tests
This change moves the rate-limiting logic from ShouldRateLimitUpload()
into a separate helper function with its own tests. The logic itself
is unchanged.

This is a precursor to changes that will experimentally change the
rate limit. The new tests will help ensure that future changes don't
unexpectedly break existing logic that should be kept (e.g., the
CrashSkippedReason, which is emitted to a histogram).

Bug: 42310127
Change-Id: I68268ee6ea76aa31961f3791c595503a204ec5b4
Reviewed-on: https://chromium-review.googlesource.com/c/crashpad/crashpad/+/7038272
Commit-Queue: Jesse McKenna <jessemckenna@google.com>
Reviewed-by: Mark Mentovai <mark@chromium.org>