Add support for attachments on macOS.
This change modifies the Crashpad client on macOS to accept a list of
file paths as attachments. These attachments are passed as arguments to
the handler process. The macOS exception handler is updated to read
these files and include them in the generated crash report.

DISALBE_SPELLCHECKER: Existing comment
Change-Id: Icf74714c76215ea61136287d1118ac3795d6f6d8
Reviewed-on: https://chromium-review.googlesource.com/c/crashpad/crashpad/+/7214128
Reviewed-by: Joshua Peraza <jperaza@chromium.org>
Reviewed-by: Ivan Penkov <ivanpe@google.com>
Commit-Queue: Yilong Yao <yilongyao@google.com>